Move logout-popin/ to ir/
