+++ components/PopinManager.jsx
@@ -11,8 +11,7 @@ import { closeCurrentPopin } from "../redux/actions";
 import AjouterCasTypes from "./ajouter-cas-types";
 import LoginForm from "./connexion";
 import EnSavoirPlus from "./en-savoir-plus";
-import Loginpopin from "./ir/login-popin";
-import LogoutPopin from "./logout-popin";
+import { LoginPopin, LogoutPopin } from "./ir";
 
 const styles = theme => ({
   dialog: {
@@ -85,7 +84,7 @@ class PopinManager extends PureComponent {
         open={showPopin}
         onClose={closeCurrentPopin}>
         <DialogContent classes={{ root: classes.dialogContent }}>
-          <Loginpopin />
+          <LoginPopin />
         </DialogContent>
       </Dialog>
     );
