+++ components/ir/index.ts
@@ -1,2 +1,4 @@
 export { default as Articles } from "./articles";
 export { default as CartesImpact } from "./cartes-impact";
+export { default as LoginPopin } from "./login-popin";
+export { default as LogoutPopin } from "./logout-popin";
