+++ components/ir/logout-popin/index.js
@@ -2,7 +2,7 @@ import { withRouter } from "next/router";
 import { connect } from "react-redux";
 import { compose } from "redux";
 
-import { closeCurrentPopin, fetchMetadataCasTypes } from "../../redux/actions";
+import { closeCurrentPopin, fetchMetadataCasTypes } from "../../../redux/actions";
 import LogoutPopinComponent from "./logout-popin-component";
 
 const mapStateToProps = null;
